+++ packages/core/test/test-util/test-services/v4/test-service/TestEntity.ts
@@ -8,6 +8,7 @@ import { Moment, Duration } from 'moment';
 import { BigNumber } from 'bignumber.js';
 import { TestComplexType, TestComplexTypeField } from './TestComplexType';
 import { TestEnumType } from './TestEnumType';
+import { TestEnumTypeInt64 } from './TestEnumTypeInt64';
 import { TestEnumTypeWithOneMember } from './TestEnumTypeWithOneMember';
 import {
   AllFields,
@@ -158,6 +159,11 @@ export class TestEntity extends EntityV4 implements TestEntityType {
    */
   enumProperty?: TestEnumType;
   /**
+   * Enum Property Int 64.
+   * @nullable
+   */
+  enumPropertyInt64?: TestEnumTypeInt64;
+  /**
    * Enum Property With One Member.
    * @nullable
    */
@@ -248,6 +254,7 @@ export interface TestEntityType {
   complexTypeProperty?: TestComplexType | null;
   complexTypeCollectionProperty?: TestComplexType[] | null;
   enumProperty?: TestEnumType | null;
+  enumPropertyInt64?: TestEnumTypeInt64 | null;
   enumPropertyWithOneMember?: TestEnumTypeWithOneMember | null;
   enumCollectionProperty?: TestEnumType[] | null;
   toMultiLink: TestEntityMultiLinkType[];
@@ -477,6 +484,15 @@ export namespace TestEntity {
     true
   );
   /**
+   * Static representation of the [[enumPropertyInt64]] property for query construction.
+   * Use to reference this property in query operations such as 'select' in the fluent request API.
+   */
+  export const ENUM_PROPERTY_INT_64 = _fieldBuilder.buildEdmTypeField(
+    'EnumPropertyInt64',
+    'Edm.Enum',
+    true
+  );
+  /**
    * Static representation of the [[enumPropertyWithOneMember]] property for query construction.
    * Use to reference this property in query operations such as 'select' in the fluent request API.
    */
@@ -569,6 +585,7 @@ export namespace TestEntity {
     TestEntity.COMPLEX_TYPE_PROPERTY,
     TestEntity.COMPLEX_TYPE_COLLECTION_PROPERTY,
     TestEntity.ENUM_PROPERTY,
+    TestEntity.ENUM_PROPERTY_INT_64,
     TestEntity.ENUM_PROPERTY_WITH_ONE_MEMBER,
     TestEntity.ENUM_COLLECTION_PROPERTY,
     TestEntity.TO_MULTI_LINK,