+++ packages/generator/src/generator.spec.ts
@@ -1,4 +1,5 @@
 import { resolve } from 'path';
+import nock = require('nock');
 import { FunctionDeclaration, SourceFile } from 'ts-morph';
 import { createOptions } from '../test/test-util/create-generator-options';
 import {
@@ -32,6 +33,7 @@ describe('generator', () => {
     });
 
     it('generates the api hub metadata and writes to the input folder', async () => {
+      nock('http://registry.npmjs.org/').head(/.*/).reply(404);
       const project = await generateProject(
         createOptions({
           inputDir: resolve(oDataServiceSpecs, 'v2', 'API_TEST_SRV'),
@@ -116,7 +118,7 @@ describe('generator', () => {
     });
 
     it('generates expected number of files', () => {
-      expect(files.length).toBe(38);
+      expect(files.length).toBe(41);
     });
 
     it('generates TestEntity.ts file', () => {
@@ -137,6 +139,7 @@ describe('generator', () => {
         'bignumber.js',
         './TestComplexType',
         './TestEnumType',
+        './TestEnumTypeInt64',
         './TestEnumTypeWithOneMember',
         '@sap-cloud-sdk/core',
         './TestEntityMultiLink',
@@ -144,12 +147,12 @@ describe('generator', () => {
       ]);
 
       const entityClass = testEntityFile!.getClass('TestEntity');
-      expect(entityClass!.getProperties().length).toBe(31);
+      expect(entityClass!.getProperties().length).toBe(32);
 
       checkStaticProperties(entityClass!);
 
       const entityNamespace = testEntityFile!.getModule('TestEntity');
-      expect(entityNamespace!.getVariableDeclarations().length).toBe(34);
+      expect(entityNamespace!.getVariableDeclarations().length).toBe(35);
     });
 
     it('generates function-imports.ts file', () => {