+++ CHANGELOG.md
@@ -19,18 +19,39 @@
 ## New Functionality
 
 - [core] Support `OAuth2JWTBearer` authentication type.
-- [core] Support additional headers and query parameters set on destinations.
+- [core] Support `OAuth2JWTBearer` authentication type.
+- [core] Support additional headers and query parameters, that are set on destinations.
+-
 
 ## Improvements
 
--
+- [odata-generator] Add underlying type and value information as part of the API documentation for enums.
+
+## Fixed Issues
+
+- [odata-generator] Fix generation errors, when Enum type field is used as a key of an entity.
+
+# 1.47.1
+
+Release Date: TBD<br>
+API Docs: https://sap.github.io/cloud-sdk/api/1.47.1<br>
+Blog: TBD<br>
+
+# 1.47.0
+
+Release Date: TBD<br>
+API Docs: https://sap.github.io/cloud-sdk/api/1.47.0<br>
+Blog: TBD<br>
+
+## New Functionality
+
+- [core] Support `OAuth2JWTBearer` authentication type.
 
 ## Fixed Issues
 
 - [proxy] Fix destination service calls using web proxies.
 - [core] Fix type error to allow `null` values in filters for nullable properties.
 - [core] Fix OData filter runtime error, when using lambda expression with operands like `or`.
-
 
 # 1.46.0